enable api key

+++ app/(backend)/controller/api-key.controller.ts
@@ -1,7 +1,7 @@
 import { ApiKeyDynamicParam, ApiKeyType, ControllerBase } from '@BackEnd/types';
 import { zfd } from 'zod-form-data';
 import * as z from 'zod';
-import { ApiHandler, ResponseSuccessMessages } from '@BackEnd/lib';
+import { ApiHandler, ResponseErrorMessages, ResponseSuccessMessages } from '@BackEnd/lib';
 import { ApiKey } from '@BackEnd/models/api-key.model';
 import { NextResponse } from 'next/server';
 import { v4 } from 'uuid';
@@ -17,7 +17,7 @@ export const ApiKeyController: ControllerBase<ApiKeyType, ApiKeyDynamicParam> = 
   delete: ApiKeyControllerDelete,
   update: ApiKeyControllerCreate,
   getAll: ApiKeyControllerCreate,
-  find: ApiKeyControllerCreate,
+  find: ApiKeyControllerFind,
 };
 
 // ApiKey Creator
@@ -38,6 +38,7 @@ async function ApiKeyControllerCreate(req: Request): Promise<NextResponse> {
       schema: createSchema,
       errorMessage: undefined,
       hasBody: true,
+      isVerify: false,
     }
   );
 }
@@ -49,10 +50,38 @@ async function ApiKeyControllerDelete(
 ): Promise<NextResponse> {
   return await ApiHandler<ApiKeyType>(
     async () => {
-      await ApiKey.deleteOne({ _id: params._id, apiKey: params.apiKey });
+      const result = await ApiKey.findOne({ _id: params._id, apiKey: params.apiKey });
+      if (result) {
+        await ApiKey.deleteOne({ _id: params._id, apiKey: params.apiKey });
+        return {
+          data: null,
+          message: ResponseSuccessMessages.DeleteApiKey,
+        };
+      } else {
+        return {
+          data: null,
+          message: ResponseErrorMessages.APIKeyIsNotAvailable,
+        };
+      }
+    },
+    {
+      req,
+      schema: createSchema,
+    }
+  );
+}
+
+// ApiKey Find
+async function ApiKeyControllerFind(
+  req: Request,
+  { params }: ApiKeyDynamicParam
+): Promise<NextResponse> {
+  return await ApiHandler<ApiKeyType>(
+    async () => {
+      const result = await ApiKey.findOne({ apiKey: params.apiKey });
       return {
-        data: null,
-        message: ResponseSuccessMessages.DeleteApiKey,
+        data: result,
+        message: ResponseSuccessMessages.FindApiKey,
       };
     },
     {
